webapp/requirements.txt
django
djangorestframework
djangorestframework-xml